refactor(core): extract SignalHandlerManager from memory.ts
Round 5 Candidate #5 — extract the process signal handler singleton
from memory.ts into its own module to isolate mutable global state
(SIGTERM/SIGINT listeners) from the memory store logic.

- signal-handler-manager.ts (new, ~40 lines) — extracted singleton
  with register/unregister API, lazy global handler installation
- memory.ts (modified, -25 lines) — imports from new module,
  removed local definition
- Uses type-only import (import type { MemoryStore }) to avoid
  circular dependency at runtime

70 memory tests pass. Lint clean (244 files).
